move draw user logic to sep function, draw to same canvas
was kind of messing around and seeing how the code works - made a couple of changes and put them in a separate branch just so that nothing would be overwritten. Not really sure what you had in mind, but I put these in the same canvas rather than having 2 separate canvases. (the second canvas was not working on my comp)