add color attribute [not entirely done]
assign each user a color and draw them on the screen. Also scale their position based on the size of the current canvas. Two problems currently - the grid is not to scale i.e. on a 16x9 display, the screen is warped. I think enforcing a square grid with blank padding on the side of the screen